upadte the starrocks update.

+++ impl/src/main/java/org/pentaho/di/trans/steps/starrockskettleconnector/StarRocksMeta.java
@@ -23,7 +23,7 @@ public interface StarRocksMeta extends StepMetaInterface {
 
     long getMaxrows();
 
-    long getConnecttimeout();
+    int getConnecttimeout();
 
     long getTimeout();
 }
